Move Validator to common module
