+++ src/main/windows/mainWindow.test.js
@@ -12,7 +12,7 @@ import Config from 'common/config';
 import {DEFAULT_WINDOW_HEIGHT, DEFAULT_WINDOW_WIDTH} from 'common/utils/constants';
 
 import ContextMenu from '../contextMenu';
-import * as Validator from '../Validator';
+import * as Validator from 'common/Validator';
 
 import createMainWindow from './mainWindow';
 
@@ -58,7 +58,7 @@ jest.mock('fs', () => ({
     writeFileSync: jest.fn(),
 }));
 
-jest.mock('../Validator', () => ({
+jest.mock('common/Validator', () => ({
     validateBoundsInfo: jest.fn(),
 }));
